+++ spec/rubocop/cop/performance/string_replacement_spec.rb
@@ -68,25 +68,37 @@ describe RuboCop::Cop::Performance::StringReplacement do
       end
 
       it 'allows empty regex literal pattern' do
-        inspect_source(cop, "'abc'.gsub(//, '1')")
+        inspect_source(cop, "'abc'.#{method}(//, '1')")
 
         expect(cop.messages).to be_empty
       end
 
       it 'allows empty regex pattern from string' do
-        inspect_source(cop, "'abc'.gsub(Regexp.new(''), '1')")
+        inspect_source(cop, "'abc'.#{method}(Regexp.new(''), '1')")
 
         expect(cop.messages).to be_empty
       end
 
       it 'allows empty regex pattern from regex' do
-        inspect_source(cop, "'abc'.gsub(Regexp.new(//), '1')")
+        inspect_source(cop, "'abc'.#{method}(Regexp.new(//), '1')")
+
+        expect(cop.messages).to be_empty
+      end
+
+      it 'allows regex literals with options' do
+        inspect_source(cop, "'abc'.#{method}(/a/i, '1')")
+
+        expect(cop.messages).to be_empty
+      end
+
+      it 'allows regex with options' do
+        inspect_source(cop, "'abc'.#{method}(Regexp.new(/a/i), '1')")
 
         expect(cop.messages).to be_empty
       end
 
       it 'allows empty string pattern' do
-        inspect_source(cop, "'abc'.gsub('', '1')")
+        inspect_source(cop, "'abc'.#{method}('', '1')")
 
         expect(cop.messages).to be_empty
       end